Rename api js module to web
