+++ lib/api/src/extern_ref.rs
@@ -3,7 +3,7 @@ use std::any::Any;
 use crate::store::{AsStoreMut, AsStoreRef};
 
 #[cfg(feature = "web")]
-use crate::js::extern_ref as extern_ref_imp;
+use crate::web::extern_ref as extern_ref_imp;
 #[cfg(feature = "jsc")]
 use crate::jsc::extern_ref as extern_ref_imp;
 #[cfg(feature = "sys")]
