+++ lib/api/src/externals/function.rs
@@ -1,5 +1,5 @@
 #[cfg(feature = "web")]
-use crate::js::externals::function as function_impl;
+use crate::web::externals::function as function_impl;
 #[cfg(feature = "jsc")]
 use crate::jsc::externals::function as function_impl;
 #[cfg(feature = "sys")]
