+++ lib/api/src/externals/global.rs
@@ -9,7 +9,7 @@ use crate::Mutability;
 use crate::RuntimeError;
 
 #[cfg(feature = "web")]
-use crate::js::externals::global as global_impl;
+use crate::web::externals::global as global_impl;
 #[cfg(feature = "jsc")]
 use crate::jsc::externals::global as global_impl;
 #[cfg(feature = "sys")]
