+++ lib/api/src/externals/memory_view.rs
@@ -5,7 +5,7 @@ use std::mem::MaybeUninit;
 use wasmer_types::Pages;
 
 #[cfg(feature = "web")]
-use crate::js::externals::memory_view as memory_view_impl;
+use crate::web::externals::memory_view as memory_view_impl;
 #[cfg(feature = "jsc")]
 use crate::jsc::externals::memory_view as memory_view_impl;
 #[cfg(feature = "sys")]
