+++ lib/api/src/module.rs
@@ -16,7 +16,7 @@ use wasmer_types::{ExportType, ImportType};
 use crate::into_bytes::IntoBytes;
 
 #[cfg(feature = "web")]
-use crate::js::module as module_imp;
+use crate::web::module as module_imp;
 #[cfg(feature = "jsc")]
 use crate::jsc::module as module_imp;
 #[cfg(feature = "sys")]
