+++ lib/api/src/web/externals/memory_view.rs
