+++ lib/api/src/web/module.rs
@@ -1,7 +1,7 @@
 use crate::errors::InstantiationError;
 use crate::errors::RuntimeError;
 use crate::imports::Imports;
-use crate::js::AsJs;
+use crate::web::AsJs;
 use crate::store::AsStoreMut;
 use crate::vm::VMInstance;
 use crate::Extern;
@@ -90,7 +90,7 @@ impl Module {
         // The module is now validated, so we can safely parse it's types
         #[cfg(feature = "wasm-types-polyfill")]
         let (type_hints, name) = {
-            let info = crate::js::module_info_polyfill::translate_module(&binary[..]).unwrap();
+            let info = crate::web::module_info_polyfill::translate_module(&binary[..]).unwrap();
 
             (
                 Some(ModuleTypeHints {
