+++ lib/api/src/web/module_info_polyfill.rs
